Move build logic to lib
